LPS-85444 extract tests to separated classes since the tests class is already too big
