+++ modules/apps/headless-apio/structured-content/structured-content-apio-test/src/testIntegration/java/com/liferay/structured/content/apio/internal/architect/resource/test/StructuredContentNestedCollectionResourcePermissionTest.java
@@ -1,0 +1,242 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.structured.content.apio.internal.architect.resource.test;
+
+import com.liferay.apio.architect.pagination.PageItems;
+import com.liferay.apio.architect.pagination.Pagination;
+import com.liferay.apio.architect.resource.NestedCollectionResource;
+import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
+import com.liferay.journal.model.JournalArticle;
+import com.liferay.journal.model.JournalArticleConstants;
+import com.liferay.journal.model.JournalArticleWrapper;
+import com.liferay.journal.model.JournalFolderConstants;
+import com.liferay.journal.service.JournalArticleLocalService;
+import com.liferay.journal.test.util.JournalTestUtil;
+import com.liferay.portal.kernel.model.Company;
+import com.liferay.portal.kernel.model.Group;
+import com.liferay.portal.kernel.model.User;
+import com.liferay.portal.kernel.security.auth.PrincipalException;
+import com.liferay.portal.kernel.security.permission.PermissionChecker;
+import com.liferay.portal.kernel.security.permission.PermissionCheckerFactoryUtil;
+import com.liferay.portal.kernel.service.CompanyLocalServiceUtil;
+import com.liferay.portal.kernel.service.ServiceContext;
+import com.liferay.portal.kernel.service.UserLocalService;
+import com.liferay.portal.kernel.test.context.ContextUserReplace;
+import com.liferay.portal.kernel.test.rule.AggregateTestRule;
+import com.liferay.portal.kernel.test.rule.DeleteAfterTestRun;
+import com.liferay.portal.kernel.test.util.GroupTestUtil;
+import com.liferay.portal.kernel.test.util.RandomTestUtil;
+import com.liferay.portal.kernel.test.util.ServiceContextTestUtil;
+import com.liferay.portal.kernel.test.util.UserTestUtil;
+import com.liferay.portal.kernel.theme.ThemeDisplay;
+import com.liferay.portal.kernel.util.LocaleUtil;
+import com.liferay.portal.test.rule.Inject;
+import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
+import com.liferay.portal.test.rule.PermissionCheckerTestRule;
+import com.liferay.structured.content.apio.architect.filter.Filter;
+import com.liferay.structured.content.apio.architect.sort.Sort;
+import com.liferay.structured.content.apio.architect.util.test.PaginationTestUtil;
+
+import java.lang.reflect.InvocationTargetException;
+import java.lang.reflect.Method;
+
+import java.util.HashMap;
+import java.util.Locale;
+import java.util.Map;
+
+import org.assertj.core.api.Assertions;
+
+import org.junit.Assert;
+import org.junit.Before;
+import org.junit.ClassRule;
+import org.junit.Rule;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+
+/**
+ * @author Cristina González
+ */
+@RunWith(Arquillian.class)
+public class StructuredContentNestedCollectionResourcePermissionTest {
+
+	@ClassRule
+	@Rule
+	public static final AggregateTestRule aggregateTestRule =
+		new AggregateTestRule(
+			new LiferayIntegrationTestRule(),
+			PermissionCheckerTestRule.INSTANCE);
+
+	@Before
+	public void setUp() throws Exception {
+		_group = GroupTestUtil.addGroup();
+	}
+
+	@Test
+	public void testGetJournalArticleWrapperFilterByPermission()
+		throws Exception {
+
+		Map<Locale, String> stringMap = new HashMap<>();
+
+		String title = RandomTestUtil.randomString();
+
+		stringMap.put(LocaleUtil.getDefault(), title);
+
+		ServiceContext serviceContext =
+			ServiceContextTestUtil.getServiceContext(_group.getGroupId());
+
+		serviceContext.setAddGuestPermissions(false);
+		serviceContext.setAddGroupPermissions(false);
+
+		JournalArticle journalArticle = JournalTestUtil.addArticle(
+			_group.getGroupId(),
+			JournalFolderConstants.DEFAULT_PARENT_FOLDER_ID,
+			JournalArticleConstants.CLASSNAME_ID_DEFAULT, title, false,
+			stringMap, stringMap, stringMap, null, LocaleUtil.getDefault(),
+			null, true, true, serviceContext);
+
+		User user = UserTestUtil.addUser();
+
+		PermissionChecker permissionChecker =
+			PermissionCheckerFactoryUtil.create(user);
+
+		try (ContextUserReplace contextUserReplace =
+				new ContextUserReplace(user, permissionChecker)) {
+
+			Assertions.assertThatThrownBy(
+				() -> _getJournalArticleWrapper(
+					journalArticle.getId(),
+					_getThemeDisplay(_group, LocaleUtil.getDefault()))
+			).isInstanceOf(
+				PrincipalException.MustHavePermission.class
+			);
+		}
+		finally {
+			_userLocalService.deleteUser(user);
+		}
+	}
+
+	@Test
+	public void testGetPageItemsFilterByPermission() throws Exception {
+		Map<Locale, String> stringMap = new HashMap<>();
+
+		stringMap.put(LocaleUtil.getDefault(), RandomTestUtil.randomString());
+
+		ServiceContext serviceContext =
+			ServiceContextTestUtil.getServiceContext(_group.getGroupId());
+
+		serviceContext.setAddGuestPermissions(false);
+		serviceContext.setAddGroupPermissions(false);
+
+		JournalTestUtil.addArticle(
+			_group.getGroupId(),
+			JournalFolderConstants.DEFAULT_PARENT_FOLDER_ID,
+			JournalArticleConstants.CLASSNAME_ID_DEFAULT,
+			RandomTestUtil.randomString(), false, stringMap, stringMap,
+			stringMap, null, LocaleUtil.getDefault(), null, true, true,
+			serviceContext);
+
+		User user = UserTestUtil.addUser();
+
+		PermissionChecker permissionChecker =
+			PermissionCheckerFactoryUtil.create(user);
+
+		try (ContextUserReplace contextUserReplace =
+				new ContextUserReplace(user, permissionChecker)) {
+
+			PageItems<JournalArticle> pageItems = _getPageItems(
+				PaginationTestUtil.of(10, 1), _group.getGroupId(),
+				_getThemeDisplay(_group, LocaleUtil.getDefault()),
+				Filter.emptyFilter(), Sort.emptySort());
+
+			Assert.assertEquals(0, pageItems.getTotalCount());
+		}
+		finally {
+			_userLocalService.deleteUser(user);
+		}
+	}
+
+	private JournalArticleWrapper _getJournalArticleWrapper(
+			long journalArticleId, ThemeDisplay themeDisplay)
+		throws Throwable {
+
+		Class<? extends NestedCollectionResource> clazz =
+			_nestedCollectionResource.getClass();
+
+		Method method = clazz.getDeclaredMethod(
+			"_getJournalArticleWrapper", long.class, ThemeDisplay.class);
+
+		method.setAccessible(true);
+
+		try {
+			return (JournalArticleWrapper)method.invoke(
+				_nestedCollectionResource, journalArticleId, themeDisplay);
+		}
+		catch (InvocationTargetException ite) {
+			ite.printStackTrace();
+
+			throw ite.getCause();
+		}
+	}
+
+	private PageItems<JournalArticle> _getPageItems(
+			Pagination pagination, long contentSpaceId,
+			ThemeDisplay themeDisplay, Filter filter, Sort sort)
+		throws Exception {
+
+		Class<? extends NestedCollectionResource> clazz =
+			_nestedCollectionResource.getClass();
+
+		Method method = clazz.getDeclaredMethod(
+			"_getPageItems", Pagination.class, long.class, ThemeDisplay.class,
+			Filter.class, Sort.class);
+
+		method.setAccessible(true);
+
+		return (PageItems)method.invoke(
+			_nestedCollectionResource, pagination, contentSpaceId, themeDisplay,
+			filter, sort);
+	}
+
+	private ThemeDisplay _getThemeDisplay(Group group, Locale locale)
+		throws Exception {
+
+		ThemeDisplay themeDisplay = new ThemeDisplay();
+
+		Company company = CompanyLocalServiceUtil.getCompanyById(
+			group.getCompanyId());
+
+		themeDisplay.setCompany(company);
+
+		themeDisplay.setLocale(locale);
+		themeDisplay.setScopeGroupId(group.getGroupId());
+
+		return themeDisplay;
+	}
+
+	@DeleteAfterTestRun
+	private Group _group;
+
+	@Inject
+	private JournalArticleLocalService _journalArticleLocalService;
+
+	@Inject(
+		filter = "component.name=com.liferay.structured.content.apio.internal.architect.resource.StructuredContentNestedCollectionResource"
+	)
+	private NestedCollectionResource _nestedCollectionResource;
+
+	@Inject
+	private UserLocalService _userLocalService;
+
+}
