+++ modules/apps/headless-apio/structured-content/structured-content-apio-test/src/testIntegration/java/com/liferay/structured/content/apio/internal/architect/resource/test/StructuredContentNestedCollectionResourceTest.java
@@ -27,11 +27,6 @@ import com.liferay.journal.test.util.JournalTestUtil;
 import com.liferay.portal.kernel.model.Company;
 import com.liferay.portal.kernel.model.Group;
 import com.liferay.portal.kernel.model.User;
-import com.liferay.portal.kernel.search.BooleanClause;
-import com.liferay.portal.kernel.search.BooleanClauseOccur;
-import com.liferay.portal.kernel.search.Query;
-import com.liferay.portal.kernel.search.QueryTerm;
-import com.liferay.portal.kernel.search.generic.TermQueryImpl;
 import com.liferay.portal.kernel.security.auth.PrincipalException;
 import com.liferay.portal.kernel.security.permission.PermissionChecker;
 import com.liferay.portal.kernel.security.permission.PermissionCheckerFactoryUtil;
@@ -52,9 +47,7 @@ import com.liferay.portal.test.rule.Inject;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
 import com.liferay.portal.test.rule.PermissionCheckerTestRule;
 import com.liferay.structured.content.apio.architect.filter.Filter;
-import com.liferay.structured.content.apio.architect.filter.FilterParser;
 import com.liferay.structured.content.apio.architect.sort.Sort;
-import com.liferay.structured.content.apio.architect.sort.SortParser;
 import com.liferay.structured.content.apio.architect.util.test.PaginationTestUtil;
 
 import java.lang.reflect.InvocationTargetException;
@@ -91,25 +84,6 @@ public class StructuredContentNestedCollectionResourceTest {
 	@Before
 	public void setUp() throws Exception {
 		_group = GroupTestUtil.addGroup();
-	}
-
-	@Test
-	public void testGetBooleanClauseWithExistingProperty() throws Exception {
-		BooleanClause<Query> queryBooleanClause = getBooleanClause(
-			new Filter(_filterParser.parse("title eq 'Title Value'")),
-			LocaleUtil.US);
-
-		Assert.assertEquals(
-			BooleanClauseOccur.MUST,
-			queryBooleanClause.getBooleanClauseOccur());
-
-		TermQueryImpl termQuery = (TermQueryImpl)queryBooleanClause.getClause();
-
-		QueryTerm queryTerm = termQuery.getQueryTerm();
-
-		Assert.assertEquals(
-			"localized_title_en_US_sortable", queryTerm.getField());
-		Assert.assertEquals("title value", queryTerm.getValue());
 	}
 
 	@Test
@@ -214,102 +188,6 @@ public class StructuredContentNestedCollectionResourceTest {
 	}
 
 	@Test
-	public void testGetPageItemsFilterByPartialTitle() throws Exception {
-		Map<Locale, String> stringMap = new HashMap<>();
-
-		stringMap.put(LocaleUtil.getDefault(), "hello world");
-
-		ServiceContext serviceContext =
-			ServiceContextTestUtil.getServiceContext(_group.getGroupId());
-
-		JournalTestUtil.addArticle(
-			_group.getGroupId(),
-			JournalFolderConstants.DEFAULT_PARENT_FOLDER_ID,
-			JournalArticleConstants.CLASSNAME_ID_DEFAULT,
-			RandomTestUtil.randomString(), false, stringMap, stringMap,
-			stringMap, null, LocaleUtil.getDefault(), null, true, true,
-			serviceContext);
-
-		PageItems<JournalArticle> pageItems = _getPageItems(
-			PaginationTestUtil.of(10, 1), _group.getGroupId(),
-			_getThemeDisplay(_group, LocaleUtil.getDefault()),
-			new Filter(_filterParser.parse("(title eq 'hello')")),
-			Sort.emptySort());
-
-		Assert.assertEquals(0, pageItems.getTotalCount());
-	}
-
-	@Test
-	public void testGetPageItemsFilterByPartialTitleWithSeveralWords()
-		throws Exception {
-
-		Map<Locale, String> stringMap1 = new HashMap<>();
-
-		stringMap1.put(LocaleUtil.getDefault(), "hello world");
-
-		ServiceContext serviceContext =
-			ServiceContextTestUtil.getServiceContext(_group.getGroupId());
-
-		JournalArticle journalArticle1 = JournalTestUtil.addArticle(
-			_group.getGroupId(),
-			JournalFolderConstants.DEFAULT_PARENT_FOLDER_ID,
-			JournalArticleConstants.CLASSNAME_ID_DEFAULT,
-			RandomTestUtil.randomString(), false, stringMap1, stringMap1,
-			stringMap1, null, LocaleUtil.getDefault(), null, true, true,
-			serviceContext);
-
-		Map<Locale, String> stringMap2 = new HashMap<>();
-
-		stringMap2.put(LocaleUtil.getDefault(), "hello");
-
-		JournalTestUtil.addArticle(
-			_group.getGroupId(),
-			JournalFolderConstants.DEFAULT_PARENT_FOLDER_ID,
-			JournalArticleConstants.CLASSNAME_ID_DEFAULT,
-			RandomTestUtil.randomString(), false, stringMap2, stringMap2,
-			stringMap2, null, LocaleUtil.getDefault(), null, true, true,
-			serviceContext);
-
-		PageItems<JournalArticle> pageItems = _getPageItems(
-			PaginationTestUtil.of(10, 1), _group.getGroupId(),
-			_getThemeDisplay(_group, LocaleUtil.getDefault()),
-			new Filter(_filterParser.parse("(title eq 'hello world')")),
-			Sort.emptySort());
-
-		Assert.assertEquals(1, pageItems.getTotalCount());
-
-		List<JournalArticle> items = (List<JournalArticle>)pageItems.getItems();
-
-		Assert.assertTrue("Items " + items, items.contains(journalArticle1));
-	}
-
-	@Test
-	public void testGetPageItemsFilterByPartialTitleWord() throws Exception {
-		Map<Locale, String> stringMap = new HashMap<>();
-
-		stringMap.put(LocaleUtil.getDefault(), "hello world");
-
-		ServiceContext serviceContext =
-			ServiceContextTestUtil.getServiceContext(_group.getGroupId());
-
-		JournalTestUtil.addArticle(
-			_group.getGroupId(),
-			JournalFolderConstants.DEFAULT_PARENT_FOLDER_ID,
-			JournalArticleConstants.CLASSNAME_ID_DEFAULT,
-			RandomTestUtil.randomString(), false, stringMap, stringMap,
-			stringMap, null, LocaleUtil.getDefault(), null, true, true,
-			serviceContext);
-
-		PageItems<JournalArticle> pageItems = _getPageItems(
-			PaginationTestUtil.of(10, 1), _group.getGroupId(),
-			_getThemeDisplay(_group, LocaleUtil.getDefault()),
-			new Filter(_filterParser.parse("(title eq 'hel')")),
-			Sort.emptySort());
-
-		Assert.assertEquals(0, pageItems.getTotalCount());
-	}
-
-	@Test
 	public void testGetPageItemsFilterByPermission() throws Exception {
 		Map<Locale, String> stringMap = new HashMap<>();
 
@@ -347,451 +225,6 @@ public class StructuredContentNestedCollectionResourceTest {
 		finally {
 			_userLocalService.deleteUser(user);
 		}
-	}
-
-	@Test
-	public void testGetPageItemsFilterByTitle() throws Exception {
-		Map<Locale, String> stringMap1 = new HashMap<>();
-
-		stringMap1.put(LocaleUtil.getDefault(), "title1");
-
-		ServiceContext serviceContext =
-			ServiceContextTestUtil.getServiceContext(_group.getGroupId());
-
-		JournalArticle journalArticle1 = JournalTestUtil.addArticle(
-			_group.getGroupId(),
-			JournalFolderConstants.DEFAULT_PARENT_FOLDER_ID,
-			JournalArticleConstants.CLASSNAME_ID_DEFAULT,
-			RandomTestUtil.randomString(), false, stringMap1, stringMap1,
-			stringMap1, null, LocaleUtil.getDefault(), null, true, true,
-			serviceContext);
-
-		Map<Locale, String> stringMap2 = new HashMap<>();
-
-		stringMap2.put(LocaleUtil.getDefault(), "title2");
-
-		JournalTestUtil.addArticle(
-			_group.getGroupId(),
-			JournalFolderConstants.DEFAULT_PARENT_FOLDER_ID,
-			JournalArticleConstants.CLASSNAME_ID_DEFAULT,
-			RandomTestUtil.randomString(), false, stringMap2, stringMap2,
-			stringMap2, null, LocaleUtil.getDefault(), null, true, true,
-			serviceContext);
-
-		PageItems<JournalArticle> pageItems = _getPageItems(
-			PaginationTestUtil.of(10, 1), _group.getGroupId(),
-			_getThemeDisplay(_group, LocaleUtil.getDefault()),
-			new Filter(_filterParser.parse("(title eq 'title1')")),
-			Sort.emptySort());
-
-		Assert.assertEquals(1, pageItems.getTotalCount());
-
-		List<JournalArticle> items = (List<JournalArticle>)pageItems.getItems();
-
-		Assert.assertEquals(journalArticle1, items.get(0));
-	}
-
-	@Test
-	public void testGetPageItemsFilterByTitleInLocaleWithDifferentDefaultLocale()
-		throws Exception {
-
-		Map<Locale, String> stringMap = new HashMap<>();
-
-		stringMap.put(LocaleUtil.SPAIN, "titulo1");
-
-		ServiceContext serviceContext =
-			ServiceContextTestUtil.getServiceContext(_group.getGroupId());
-
-		// Default locale of article is in Spanish (no English content or title)
-
-		JournalArticle journalArticle = JournalTestUtil.addArticle(
-			_group.getGroupId(),
-			JournalFolderConstants.DEFAULT_PARENT_FOLDER_ID,
-			JournalArticleConstants.CLASSNAME_ID_DEFAULT,
-			RandomTestUtil.randomString(), false, stringMap, stringMap,
-			stringMap, null, LocaleUtil.SPAIN, null, true, true,
-			serviceContext);
-
-		PageItems<JournalArticle> pageItems = _getPageItems(
-			PaginationTestUtil.of(10, 1), _group.getGroupId(),
-			_getThemeDisplay(_group, LocaleUtil.getDefault()),
-			new Filter(_filterParser.parse("(title eq 'titulo1')")),
-			Sort.emptySort());
-
-		Assert.assertEquals(1, pageItems.getTotalCount());
-
-		List<JournalArticle> items = (List<JournalArticle>)pageItems.getItems();
-
-		Assert.assertEquals(journalArticle, items.get(0));
-	}
-
-	@Test
-	public void testGetPageItemsFilterByTitleInLocaleWithDifferentLocale()
-		throws Exception {
-
-		Map<Locale, String> stringMap = new HashMap<>();
-
-		stringMap.put(LocaleUtil.getDefault(), RandomTestUtil.randomString());
-		stringMap.put(LocaleUtil.SPAIN, "titulo1");
-
-		ServiceContext serviceContext =
-			ServiceContextTestUtil.getServiceContext(_group.getGroupId());
-
-		JournalTestUtil.addArticle(
-			_group.getGroupId(),
-			JournalFolderConstants.DEFAULT_PARENT_FOLDER_ID,
-			JournalArticleConstants.CLASSNAME_ID_DEFAULT,
-			RandomTestUtil.randomString(), false, stringMap, stringMap,
-			stringMap, null, LocaleUtil.getDefault(), null, true, true,
-			serviceContext);
-
-		PageItems<JournalArticle> pageItems = _getPageItems(
-			PaginationTestUtil.of(10, 1), _group.getGroupId(),
-			_getThemeDisplay(_group, LocaleUtil.getDefault()),
-			new Filter(_filterParser.parse("(title eq 'titulo1')")),
-			Sort.emptySort());
-
-		Assert.assertEquals(0, pageItems.getTotalCount());
-	}
-
-	@Test
-	public void testGetPageItemsFilterByTitleInLocaleWithSameLocale()
-		throws Exception {
-
-		Map<Locale, String> stringMap = new HashMap<>();
-
-		stringMap.put(LocaleUtil.getDefault(), RandomTestUtil.randomString());
-		stringMap.put(LocaleUtil.SPAIN, "titulo1");
-
-		ServiceContext serviceContext =
-			ServiceContextTestUtil.getServiceContext(_group.getGroupId());
-
-		JournalArticle journalArticle = JournalTestUtil.addArticle(
-			_group.getGroupId(),
-			JournalFolderConstants.DEFAULT_PARENT_FOLDER_ID,
-			JournalArticleConstants.CLASSNAME_ID_DEFAULT,
-			RandomTestUtil.randomString(), false, stringMap, stringMap,
-			stringMap, null, LocaleUtil.getDefault(), null, true, true,
-			serviceContext);
-
-		PageItems<JournalArticle> pageItems = _getPageItems(
-			PaginationTestUtil.of(10, 1), _group.getGroupId(),
-			_getThemeDisplay(_group, LocaleUtil.SPAIN),
-			new Filter(_filterParser.parse("(title eq 'titulo1')")),
-			Sort.emptySort());
-
-		Assert.assertEquals(1, pageItems.getTotalCount());
-
-		List<JournalArticle> items = (List<JournalArticle>)pageItems.getItems();
-
-		Assert.assertEquals(journalArticle, items.get(0));
-	}
-
-	@Test
-	public void testGetPageItemsFilterByTitleStartingWithLowerCase()
-		throws Exception {
-
-		Map<Locale, String> stringMap1 = new HashMap<>();
-
-		stringMap1.put(LocaleUtil.getDefault(), "great title");
-
-		ServiceContext serviceContext =
-			ServiceContextTestUtil.getServiceContext(_group.getGroupId());
-
-		JournalArticle journalArticle1 = JournalTestUtil.addArticle(
-			_group.getGroupId(),
-			JournalFolderConstants.DEFAULT_PARENT_FOLDER_ID,
-			JournalArticleConstants.CLASSNAME_ID_DEFAULT,
-			RandomTestUtil.randomString(), false, stringMap1, stringMap1,
-			stringMap1, null, LocaleUtil.getDefault(), null, true, true,
-			serviceContext);
-
-		Map<Locale, String> stringMap2 = new HashMap<>();
-
-		stringMap2.put(LocaleUtil.getDefault(), "Great Title");
-
-		JournalArticle journalArticle2 = JournalTestUtil.addArticle(
-			_group.getGroupId(),
-			JournalFolderConstants.DEFAULT_PARENT_FOLDER_ID,
-			JournalArticleConstants.CLASSNAME_ID_DEFAULT,
-			RandomTestUtil.randomString(), false, stringMap2, stringMap2,
-			stringMap2, null, LocaleUtil.getDefault(), null, true, true,
-			serviceContext);
-
-		PageItems<JournalArticle> pageItems = _getPageItems(
-			PaginationTestUtil.of(10, 1), _group.getGroupId(),
-			_getThemeDisplay(_group, LocaleUtil.getDefault()),
-			new Filter(_filterParser.parse("(title eq 'great title')")),
-			Sort.emptySort());
-
-		Assert.assertEquals(2, pageItems.getTotalCount());
-
-		List<JournalArticle> items = (List<JournalArticle>)pageItems.getItems();
-
-		Assert.assertTrue(items.contains(journalArticle1));
-		Assert.assertTrue(items.contains(journalArticle2));
-	}
-
-	@Test
-	public void testGetPageItemsFilterByTitleStartingWithUpperCase()
-		throws Exception {
-
-		Map<Locale, String> stringMap1 = new HashMap<>();
-
-		stringMap1.put(LocaleUtil.getDefault(), "great title");
-
-		ServiceContext serviceContext =
-			ServiceContextTestUtil.getServiceContext(_group.getGroupId());
-
-		JournalArticle journalArticle1 = JournalTestUtil.addArticle(
-			_group.getGroupId(),
-			JournalFolderConstants.DEFAULT_PARENT_FOLDER_ID,
-			JournalArticleConstants.CLASSNAME_ID_DEFAULT,
-			RandomTestUtil.randomString(), false, stringMap1, stringMap1,
-			stringMap1, null, LocaleUtil.getDefault(), null, true, true,
-			serviceContext);
-
-		Map<Locale, String> stringMap2 = new HashMap<>();
-
-		stringMap2.put(LocaleUtil.getDefault(), "Great Title");
-
-		JournalArticle journalArticle2 = JournalTestUtil.addArticle(
-			_group.getGroupId(),
-			JournalFolderConstants.DEFAULT_PARENT_FOLDER_ID,
-			JournalArticleConstants.CLASSNAME_ID_DEFAULT,
-			RandomTestUtil.randomString(), false, stringMap2, stringMap2,
-			stringMap2, null, LocaleUtil.getDefault(), null, true, true,
-			serviceContext);
-
-		PageItems<JournalArticle> pageItems = _getPageItems(
-			PaginationTestUtil.of(10, 1), _group.getGroupId(),
-			_getThemeDisplay(_group, LocaleUtil.getDefault()),
-			new Filter(_filterParser.parse("(title eq 'Great Title')")),
-			Sort.emptySort());
-
-		Assert.assertEquals(2, pageItems.getTotalCount());
-
-		List<JournalArticle> items = (List<JournalArticle>)pageItems.getItems();
-
-		Assert.assertTrue(items.contains(journalArticle1));
-		Assert.assertTrue(items.contains(journalArticle2));
-	}
-
-	@Test
-	public void testGetPageItemsSortByTitleAsc() throws Exception {
-		Map<Locale, String> stringMap1 = new HashMap<>();
-
-		stringMap1.put(LocaleUtil.getDefault(), "title1");
-
-		ServiceContext serviceContext =
-			ServiceContextTestUtil.getServiceContext(_group.getGroupId());
-
-		JournalArticle journalArticle1 = JournalTestUtil.addArticle(
-			_group.getGroupId(),
-			JournalFolderConstants.DEFAULT_PARENT_FOLDER_ID,
-			JournalArticleConstants.CLASSNAME_ID_DEFAULT,
-			RandomTestUtil.randomString(), false, stringMap1, stringMap1,
-			stringMap1, null, LocaleUtil.getDefault(), null, true, true,
-			serviceContext);
-
-		Map<Locale, String> stringMap2 = new HashMap<>();
-
-		stringMap2.put(LocaleUtil.getDefault(), "title2");
-
-		JournalArticle journalArticle2 = JournalTestUtil.addArticle(
-			_group.getGroupId(),
-			JournalFolderConstants.DEFAULT_PARENT_FOLDER_ID,
-			JournalArticleConstants.CLASSNAME_ID_DEFAULT,
-			RandomTestUtil.randomString(), false, stringMap2, stringMap2,
-			stringMap2, null, LocaleUtil.getDefault(), null, true, true,
-			serviceContext);
-
-		PageItems<JournalArticle> pageItems = _getPageItems(
-			PaginationTestUtil.of(10, 1), _group.getGroupId(),
-			_getThemeDisplay(_group, LocaleUtil.getDefault()),
-			Filter.emptyFilter(), new Sort(_sortParser.parse("title:asc")));
-
-		Assert.assertEquals(2, pageItems.getTotalCount());
-
-		List<JournalArticle> items = (List<JournalArticle>)pageItems.getItems();
-
-		Assert.assertEquals(journalArticle1, items.get(0));
-		Assert.assertEquals(journalArticle2, items.get(1));
-	}
-
-	@Test
-	public void testGetPageItemsSortByTitleAscWithDifferentDefaultLocale()
-		throws Exception {
-
-		Map<Locale, String> stringMap1 = new HashMap<>();
-
-		stringMap1.put(LocaleUtil.getDefault(), "titleZ");
-		stringMap1.put(LocaleUtil.SPAIN, "title1");
-
-		ServiceContext serviceContext =
-			ServiceContextTestUtil.getServiceContext(_group.getGroupId());
-
-		JournalArticle journalArticle1 = JournalTestUtil.addArticle(
-			_group.getGroupId(),
-			JournalFolderConstants.DEFAULT_PARENT_FOLDER_ID,
-			JournalArticleConstants.CLASSNAME_ID_DEFAULT,
-			RandomTestUtil.randomString(), false, stringMap1, stringMap1,
-			stringMap1, null, LocaleUtil.getDefault(), null, true, true,
-			serviceContext);
-
-		Map<Locale, String> stringMap2 = new HashMap<>();
-
-		stringMap2.put(LocaleUtil.SPAIN, "titleA");
-
-		JournalArticle journalArticle2 = JournalTestUtil.addArticle(
-			_group.getGroupId(),
-			JournalFolderConstants.DEFAULT_PARENT_FOLDER_ID,
-			JournalArticleConstants.CLASSNAME_ID_DEFAULT,
-			RandomTestUtil.randomString(), false, stringMap2, stringMap2,
-			stringMap2, null, LocaleUtil.SPAIN, null, true, true,
-			serviceContext);
-
-		PageItems<JournalArticle> pageItems = _getPageItems(
-			PaginationTestUtil.of(10, 1), _group.getGroupId(),
-			_getThemeDisplay(_group, LocaleUtil.getDefault()),
-			Filter.emptyFilter(), new Sort(_sortParser.parse("title:asc")));
-
-		Assert.assertEquals(2, pageItems.getTotalCount());
-
-		List<JournalArticle> items = (List<JournalArticle>)pageItems.getItems();
-
-		Assert.assertEquals(journalArticle2, items.get(0));
-		Assert.assertEquals(journalArticle1, items.get(1));
-	}
-
-	@Test
-	public void testGetPageItemsSortByTitleAscWithNonedefaultLocale()
-		throws Exception {
-
-		Map<Locale, String> stringMap1 = new HashMap<>();
-
-		stringMap1.put(LocaleUtil.getDefault(), "title1");
-		stringMap1.put(LocaleUtil.SPAIN, "titleZ");
-
-		ServiceContext serviceContext =
-			ServiceContextTestUtil.getServiceContext(_group.getGroupId());
-
-		JournalArticle journalArticle1 = JournalTestUtil.addArticle(
-			_group.getGroupId(),
-			JournalFolderConstants.DEFAULT_PARENT_FOLDER_ID,
-			JournalArticleConstants.CLASSNAME_ID_DEFAULT,
-			RandomTestUtil.randomString(), false, stringMap1, stringMap1,
-			stringMap1, null, LocaleUtil.getDefault(), null, true, true,
-			serviceContext);
-
-		Map<Locale, String> stringMap2 = new HashMap<>();
-
-		stringMap2.put(LocaleUtil.getDefault(), "title2");
-		stringMap2.put(LocaleUtil.SPAIN, "titleA");
-
-		JournalArticle journalArticle2 = JournalTestUtil.addArticle(
-			_group.getGroupId(),
-			JournalFolderConstants.DEFAULT_PARENT_FOLDER_ID,
-			JournalArticleConstants.CLASSNAME_ID_DEFAULT,
-			RandomTestUtil.randomString(), false, stringMap2, stringMap2,
-			stringMap2, null, LocaleUtil.getDefault(), null, true, true,
-			serviceContext);
-
-		PageItems<JournalArticle> pageItems = _getPageItems(
-			PaginationTestUtil.of(10, 1), _group.getGroupId(),
-			_getThemeDisplay(_group, LocaleUtil.SPAIN), Filter.emptyFilter(),
-			new Sort(_sortParser.parse("title:asc")));
-
-		Assert.assertEquals(2, pageItems.getTotalCount());
-
-		List<JournalArticle> items = (List<JournalArticle>)pageItems.getItems();
-
-		Assert.assertEquals(journalArticle2, items.get(0));
-		Assert.assertEquals(journalArticle1, items.get(1));
-	}
-
-	@Test
-	public void testGetPageItemsSortByTitleDefault() throws Exception {
-		Map<Locale, String> stringMap1 = new HashMap<>();
-
-		stringMap1.put(LocaleUtil.getDefault(), "title B");
-
-		ServiceContext serviceContext =
-			ServiceContextTestUtil.getServiceContext(_group.getGroupId());
-
-		JournalArticle journalArticle1 = JournalTestUtil.addArticle(
-			_group.getGroupId(),
-			JournalFolderConstants.DEFAULT_PARENT_FOLDER_ID,
-			JournalArticleConstants.CLASSNAME_ID_DEFAULT,
-			RandomTestUtil.randomString(), false, stringMap1, stringMap1,
-			stringMap1, null, LocaleUtil.getDefault(), null, true, true,
-			serviceContext);
-
-		Map<Locale, String> stringMap2 = new HashMap<>();
-
-		stringMap2.put(LocaleUtil.getDefault(), "title A");
-
-		JournalArticle journalArticle2 = JournalTestUtil.addArticle(
-			_group.getGroupId(),
-			JournalFolderConstants.DEFAULT_PARENT_FOLDER_ID,
-			JournalArticleConstants.CLASSNAME_ID_DEFAULT,
-			RandomTestUtil.randomString(), false, stringMap2, stringMap2,
-			stringMap2, null, LocaleUtil.getDefault(), null, true, true,
-			serviceContext);
-
-		PageItems<JournalArticle> pageItems = _getPageItems(
-			PaginationTestUtil.of(10, 1), _group.getGroupId(),
-			_getThemeDisplay(_group, LocaleUtil.getDefault()),
-			Filter.emptyFilter(), new Sort(_sortParser.parse("title")));
-
-		Assert.assertEquals(2, pageItems.getTotalCount());
-
-		List<JournalArticle> items = (List<JournalArticle>)pageItems.getItems();
-
-		Assert.assertEquals(journalArticle2, items.get(0));
-		Assert.assertEquals(journalArticle1, items.get(1));
-	}
-
-	@Test
-	public void testGetPageItemsSortByTitleDesc() throws Exception {
-		Map<Locale, String> stringMap1 = new HashMap<>();
-
-		stringMap1.put(LocaleUtil.getDefault(), "title A");
-
-		ServiceContext serviceContext =
-			ServiceContextTestUtil.getServiceContext(_group.getGroupId());
-
-		JournalArticle journalArticle1 = JournalTestUtil.addArticle(
-			_group.getGroupId(),
-			JournalFolderConstants.DEFAULT_PARENT_FOLDER_ID,
-			JournalArticleConstants.CLASSNAME_ID_DEFAULT,
-			RandomTestUtil.randomString(), false, stringMap1, stringMap1,
-			stringMap1, null, LocaleUtil.getDefault(), null, true, true,
-			serviceContext);
-
-		Map<Locale, String> stringMap2 = new HashMap<>();
-
-		stringMap2.put(LocaleUtil.getDefault(), "title B");
-
-		JournalArticle journalArticle2 = JournalTestUtil.addArticle(
-			_group.getGroupId(),
-			JournalFolderConstants.DEFAULT_PARENT_FOLDER_ID,
-			JournalArticleConstants.CLASSNAME_ID_DEFAULT,
-			RandomTestUtil.randomString(), false, stringMap2, stringMap2,
-			stringMap2, null, LocaleUtil.getDefault(), null, true, true,
-			serviceContext);
-
-		PageItems<JournalArticle> pageItems = _getPageItems(
-			PaginationTestUtil.of(10, 1), _group.getGroupId(),
-			_getThemeDisplay(_group, LocaleUtil.getDefault()),
-			Filter.emptyFilter(), new Sort(_sortParser.parse("title:desc")));
-
-		Assert.assertEquals(2, pageItems.getTotalCount());
-
-		List<JournalArticle> items = (List<JournalArticle>)pageItems.getItems();
-
-		Assert.assertEquals(journalArticle2, items.get(0));
-		Assert.assertEquals(journalArticle1, items.get(1));
 	}
 
 	@Test
@@ -1045,22 +478,6 @@ public class StructuredContentNestedCollectionResourceTest {
 		Assert.assertEquals(0, pageItems.getTotalCount());
 	}
 
-	protected BooleanClause<Query> getBooleanClause(
-			Filter filter, Locale locale)
-		throws Exception {
-
-		Class<? extends NestedCollectionResource> clazz =
-			_nestedCollectionResource.getClass();
-
-		Method method = clazz.getDeclaredMethod(
-			"getBooleanClause", Filter.class, Locale.class);
-
-		method.setAccessible(true);
-
-		return (BooleanClause<Query>)method.invoke(
-			_nestedCollectionResource, filter, locale);
-	}
-
 	private JournalArticleWrapper _getJournalArticleWrapper(
 			long journalArticleId, ThemeDisplay themeDisplay)
 		throws Throwable {
@@ -1119,9 +536,6 @@ public class StructuredContentNestedCollectionResourceTest {
 		return themeDisplay;
 	}
 
-	@Inject
-	private FilterParser _filterParser;
-
 	@DeleteAfterTestRun
 	private Group _group;
 
@@ -1132,9 +546,6 @@ public class StructuredContentNestedCollectionResourceTest {
 		filter = "component.name=com.liferay.structured.content.apio.internal.architect.resource.StructuredContentNestedCollectionResource"
 	)
 	private NestedCollectionResource _nestedCollectionResource;
-
-	@Inject
-	private SortParser _sortParser;
 
 	@Inject
 	private UserLocalService _userLocalService;
